Add 'По окну' button with fuzzy MHOOK file matching and auto-start emulation
- Added IDC_BUTTON_LOAD_BY_WINDOW button to load MHOOK files by active window title
- Implemented fuzzy matching algorithm (filters special chars, prefix matching)
- Added auto-start emulation after file load via 'По окну'
- Fixed C4244/C4267 warnings (LRESULT/size_t to int conversions)
- Fixed unused label warning (nomove6) in HookProc.cpp
- Updated resource.rc for VS2022 compatibility (winres.h)
- Added EmbeddedSettings.bin and Gamepad support files

+++ CircleWindow.cpp
@@ -29,7 +29,7 @@ LRESULT CALLBACK CircleTranspWndProc(HWND hwnd,
 			EndPaint(hwnd,&ps);
 			break;
 		case WM_USER_MOVEWINDOW:
-			 MoveWindow(hwnd,wparam,lparam,100,100,FALSE);
+			 MoveWindow(hwnd, static_cast<int>(wparam), static_cast<int>(lparam), 100, 100, FALSE);
 			 break;
 		case WM_CLOSE: // Çàêðûòèå îêíà èãíîðèðóåòñÿ
 			break;
